Added support for compiling resource file associated with a stylesheet
This allows .qss file to use links like ':/qss_icons/rc/radio_checked.png'.
Whenever user attempts to load a .qss file the program will try to search
and compile a .qrc file of the same name and register compiled binary as a
resource so that the images that the .qss refers to would work correctly.